pyproject.toml
[tool.poetry]
name = "web-apis-with-python"
version = "0.1.0"
description = ""
authors = ["Egor Krapivin <egor.krapivin98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
backcall = "0.2.0"
click = "8.0.1"
decorator = "5.0.9"
Flask = "2.0.1"
gunicorn = "20.1.0"
ipykernel = "5.5.5"
ipython = "7.24.0"
ipython_genutils = "0.2.0"
itsdangerous = "2.0.1"
jedi = "0.18.0"
Jinja2 = "3.0.1"
jupyter-client = "6.1.12"
jupyter-core = "4.7.1"
MarkupSafe = "2.0.1"
matplotlib-inline = "0.1.2"
parso = "0.8.2"
pexpect = "4.8.0"
pickleshare = "0.7.5"
Pillow = "9.0.0"
prompt-toolkit = "3.0.18"
ptyprocess = "0.7.0"
Pygments = "2.9.0"
python-dateutil = "2.8.1"
pyzmq = "22.1.0"
six = "1.16.0"
tornado = "6.1"
traitlets = "5.0.5"
wcwidth = "0.2.5"
Werkzeug = "2.0.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
